requirements.txt
langchain==0.1.14
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.41.1
# llama-cpp-python==0.2.53
PyMuPDF==1.23.22
python-docx==1.1.0
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
tqdm==4.66.4